card_bin_ex v1.0.0 CardBinEx

Module to discover the CreditCard brand based in the first digits.

Valid CreditCard brands

brand	alias

Visa	visa

Elo	elo

JBC	jbc

Diners	diners

Discover	discover

MasterCard	master

American Express	amex

Functions

brand_from_number(number)

Get the creditcard brand base in the first digits.

Examples

iex> CardBinEx.brand_from_number("4716892")
{:ok, "visa"}

iex> CardBinEx.brand_from_number("9716892")
{:error, :card_brand, "9716892"}

brand_from_number!(number)

Get the creditcard brand base in the first digits and return only the brand.

Examples

  iex> CardBinEx.brand_from_number!("4716892")
  "visa"

  iex> try do
  ...>   CardBinEx.brand_from_number!("9716892")
  ...> rescue
  ...>   e in CardBinEx.Error -> IO.puts(e.message)
  ...> end
  invalid card number